Move send email to helper function
