+++ backend/src/users/controllers.ts
@@ -6,8 +6,7 @@ import admin from "firebase-admin";
 import prisma from "../../client";
 import { setVolunteerCustomClaims } from "../middleware/auth";
 import userController from "../users/controllers";
-import sgMail from "@sendgrid/mail";
-import { readFile } from "fs/promises";
+import { sendEmail } from "../utils/helpers";
 
 const htmlRegCancel = "./src/emails/Registration_Cancellation.html";
 const htmlRegSuccess = "./src/emails/Registration_Successful.html";
@@ -628,34 +627,6 @@ const deleteUser = async (userID: string) => {
       id: userID,
     },
   });
-};
-
-/**
- * Sends an email to the specified address
- * @param email is the email address to send to
- * @param subject is the email subject
- * @param path is the path of the email template
- */
-const sendEmail = async (email: string, subject: string, path: string) => {
-  try {
-    // Loads an email template
-    const html = await readFile(path, "utf-8");
-    console.log("File content:", html);
-
-    // Create an email message
-    const msg = {
-      to: email, // Recipient's email address
-      from: "lagosfoodbankdev@gmail.com", // Sender's email address
-      subject: subject, // Email subject
-      html: html, // HTML body content
-    };
-
-    // Send the email
-    await sgMail.send(msg);
-    console.log("Email sent successfully!");
-  } catch (err) {
-    console.error("Error sending email:", err);
-  }
 };
 
 export default {
