+++ backend/src/utils/helpers.ts
@@ -6,6 +6,8 @@ import deleteUser from "../users/controllers";
 import prisma from "../../client";
 import { WebSocket } from "ws";
 import { wss } from "../server";
+import sgMail from "@sendgrid/mail";
+import { readFile } from "fs/promises";
 
 /**
  * Attempts the given controller and sends a success code or error code as necessary
@@ -88,4 +90,36 @@ export const socketNotify = (resource: string) => {
       client.send(JSON.stringify(dataToSend));
     }
   });
+};
+
+/**
+ * Sends an email to the specified address
+ * @param email is the email address to send to
+ * @param subject is the email subject
+ * @param path is the path of the email template
+ */
+export const sendEmail = async (
+  email: string,
+  subject: string,
+  path: string
+) => {
+  try {
+    // Loads an email template
+    const html = await readFile(path, "utf-8");
+    console.log("File content:", html);
+
+    // Create an email message
+    const msg = {
+      to: email, // Recipient's email address
+      from: "lagosfoodbankdev@gmail.com", // Sender's email address
+      subject: subject, // Email subject
+      html: html, // HTML body content
+    };
+
+    // Send the email
+    await sgMail.send(msg);
+    console.log("Email sent successfully!");
+  } catch (err) {
+    console.error("Error sending email:", err);
+  }
 };
